requirements.txt
RPi.GPIO
Seeed-grove.py
seeed-python-mlx9064x
line-bot-sdk